+++ common/model/src/main/java/is/codion/common/model/preferences/FilePreferences.java
@@ -145,4 +145,8 @@ final class FilePreferences extends AbstractPreferences {
 			throw new BackingStoreException(e);
 		}
 	}
+
+	void delete() throws IOException {
+		store.delete();
+	}
 }